Update README.md
The 'transport' setting in the configuration can only have either a 'dsn' field or the other fields, but not both.  Also the 'schema' => 'smpts' field is misleading.  The 'port' => '465' field is what turns on SSL.  The 'smpts' schema causes the underlying code to use a different Transport class.  The default class would work better in most cases.